Fix resolver dependency comparison
Our underlying resolver, Molinillo, is prepared to group spec groups
that all have the same dependencies to simplify resolution. However, our
dependency proxies were not properly comparable between each other and
these groups were not being properly built, causing candidates with the
same dependencies to be tried independently.

This was actually the cause of a very slow Gemfile, and I had added a
fix that prevented fallback groups from being added if the dependencies
of the ruby variants were the same as the platform specific variants.
Although that fix worked for that case, it's not a proper fix in general
as proved by the "infinite resolution" that's being fixed here.

The proper fix is to ignore platforms when comparing dependency proxy
objects. Platforms are only useful _outside_ of Molinillo for searching
valid candidate specifications for a dependency according to its
platform, but inside Molinillo, two `DepProxy` objects with the same
dependencies should be considered the same and grouped together even if
they have different `__platforms`.